requirements-board.txt
# Requirements for Chessnut Pro board connectivity
bleak>=0.21.1          # Bluetooth Low Energy library (macOS compatible)
python-chess>=1.10.0   # Chess logic
